Fix API tests that required permissions by using token access
