+++ ployst/core/accounts/test/test_api.py
@@ -3,27 +3,29 @@ import json
 from django.core.urlresolvers import reverse
 from django.test import TestCase
 
-from .factories import ProjectFactory, SettingsFactory, TeamFactory
+from ployst.apibase.test.mixins import CoreApiClientTestMixin
 
+from .factories import ProjectFactory, SettingsFactory, TeamFactory
 from ..models import ProjectProviderSettings
 
 TEST_TEAM = '8d7de2c4-4849-452f-82af-e142641c4b6d'
 
 
-class TestTeams(TestCase):
+class TestTeams(CoreApiClientTestMixin, TestCase):
     fixtures = ['users.json', 'teams.json', 'teamusers.json']
 
     def test_get_team_by_guid(self):
         "Test we can get a team by id"
+
         url = reverse('core:accounts:team-detail', args=[TEST_TEAM])
 
-        response = self.client.get(url)
+        response = self.client.get(url, **self.get_token_headers())
 
         team = json.loads(response.content)
         self.assertEquals(team['name'], 'Pretenders')
 
 
-class TestProjects(TestCase):
+class TestProjects(CoreApiClientTestMixin, TestCase):
 
     def test_get_projects_by_team_guid(self):
         "Test we can get a team by id"
@@ -34,7 +36,8 @@ class TestProjects(TestCase):
         project2 = ProjectFactory(name='Project Two', team=team2)
         url = reverse('core:accounts:project-list')
 
-        response = self.client.get('{}?team={}'.format(url, TEST_TEAM))
+        response = self.client.get('{}?team={}'.format(url, TEST_TEAM),
+                                   **self.get_token_headers())
 
         self.assertEquals(response.status_code, 200)
         projects = json.loads(response.content)
@@ -42,7 +45,7 @@ class TestProjects(TestCase):
         self.assertEquals(projects[0]['name'], project2.name)
 
 
-class TestProjectProviderSettings(TestCase):
+class TestProjectProviderSettings(CoreApiClientTestMixin, TestCase):
 
     def test_get_settings_by_provider_and_project(self):
         team1 = TeamFactory()
@@ -52,7 +55,8 @@ class TestProjectProviderSettings(TestCase):
         url = reverse('core:accounts:projectprovidersettings-list')
 
         response = self.client.get(
-            '{}?project={}&provider=MyProvider'.format(url, project1.id)
+            '{}?project={}&provider=MyProvider'.format(url, project1.id),
+            **self.get_token_headers()
         )
 
         self.assertEquals(response.status_code, 200)
@@ -68,7 +72,7 @@ class TestProjectProviderSettings(TestCase):
             "project": project1.id,
             "provider": 'GitHub',
             "settings": json.dumps({'c': 3})
-        })
+        }, **self.get_token_headers())
 
         settings_obj = ProjectProviderSettings.objects.all()[0]
         settings = json.loads(settings_obj.settings)
